feat: 미션관리 총 노출

+++ src/main/java/cmc/bobpossible/member/entity/Member.java
@@ -225,7 +225,10 @@ public class Member extends BaseEntity {
             store.delete();
         }
         reviews.forEach(Review::delete);
-        reward.delete();
+        if (reward != null) {
+            reward.delete();
+        }
+
     }
 
     public void addReview(Review review) {
